Change the way used to specify samesite attributes of cookies
